Add new `--ponymsgstilmute` argument to control throttling/muting
This commit add a new command line argument to control the speed of
throttling/muting called `--ponymsgstilmute`. This option is
specifically added for power users who want to have more control
over how the runtime handles backpressure. Using a value of `1`
for `--ponymsgstilmute` will result in actors getting muted after
sending a single message to a muted/overloaded/under pressure
actor and be functionally equivalent to the previous backpressure
implementation that did not employ gradual throttling.

This commit also includes some code cleanup and a minor bugfix. The
code cleanup takes the formula for determining the throttled batch
and puts it in a static function (which I'm assuming will get
inlined) to avoid potential inconsistencies between the different
places it is used. The minor bugfix is to increment
`extra_throttled_msgs_sent` in `ponyint_maybe_throttle` as the
comment indicated should be happening (instead of
`mute_map_count`).

+++ src/libponyrt/sched/scheduler.c
@@ -1312,7 +1312,7 @@ bool ponyint_sched_unmute_senders(pony_ctx_t* ctx, pony_actor_t* actor)
     while((muted = ponyint_muteset_next(&mref->value, &i)) != NULL)
     {
       // This is safe because it's a single atomic operation
-      size_t mute_map_count = atomic_fetch_sub_explicit(&muted->mute_map_count,
+      uint32_t mute_map_count = atomic_fetch_sub_explicit(&muted->mute_map_count,
         1, memory_order_relaxed);
       pony_assert(mute_map_count >= 1);
       (void)mute_map_count;
